XP-1156 Move some form json classes to admin - Moved json classes to com.enonic.xp.admin.impl.json.form" in admin-impl module - There were no tests to move - Seems like all-input-types input type should be removed from features module
