+++ modules/admin-impl/src/main/java/com/enonic/xp/admin/impl/json/application/ApplicationJson.java
@@ -6,8 +6,8 @@ import java.util.List;
 import com.google.common.collect.ImmutableList;
 
 import com.enonic.xp.admin.impl.json.ItemJson;
+import com.enonic.xp.admin.impl.json.form.FormJson;
 import com.enonic.xp.app.Application;
-import com.enonic.xp.form.FormJson;
 import com.enonic.xp.schema.mixin.MixinName;
 import com.enonic.xp.site.SiteDescriptor;
