+++ modules/admin-impl/src/main/java/com/enonic/xp/admin/impl/json/content/page/DescriptorJson.java
@@ -3,7 +3,7 @@ package com.enonic.xp.admin.impl.json.content.page;
 import com.google.common.base.Preconditions;
 
 import com.enonic.xp.admin.impl.json.ItemJson;
-import com.enonic.xp.form.FormJson;
+import com.enonic.xp.admin.impl.json.form.FormJson;
 import com.enonic.xp.region.Descriptor;
