+++ modules/admin-impl/src/main/java/com/enonic/xp/admin/impl/json/form/FieldSetJson.java
@@ -1,10 +1,14 @@
-package com.enonic.xp.form;
+package com.enonic.xp.admin.impl.json.form;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.google.common.annotations.Beta;
+
+import com.enonic.xp.form.FieldSet;
+import com.enonic.xp.form.FormItem;
+import com.enonic.xp.form.FormItems;
 
 @Beta
 @SuppressWarnings("UnusedDeclaration")
