+++ modules/admin-impl/src/main/java/com/enonic/xp/admin/impl/json/form/InlineMixinJson.java
@@ -1,7 +1,10 @@
-package com.enonic.xp.form;
+package com.enonic.xp.admin.impl.json.form;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.google.common.annotations.Beta;
+
+import com.enonic.xp.form.FormItem;
+import com.enonic.xp.form.InlineMixin;
 
 @Beta
 @SuppressWarnings("UnusedDeclaration")
