+++ modules/admin-impl/src/main/java/com/enonic/xp/admin/impl/json/form/OccurrencesJson.java
@@ -1,7 +1,9 @@
-package com.enonic.xp.form;
+package com.enonic.xp.admin.impl.json.form;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.google.common.annotations.Beta;
+
+import com.enonic.xp.form.Occurrences;
 
 @Beta
 @SuppressWarnings("UnusedDeclaration")
